+++ examples/location-retrieval.ts
@@ -16,8 +16,8 @@ const myDevice = client.devices.get({
         publicPort: 80,
     },
     Ipv6Address: "2041:0000:140F::875B:131B",
-    // The phone number accepts the "+" sign, but not spaces or "()" marks
-    phoneNumber: "36721601234567"
+    // The phone number does not accept spaces or parentheses
+    phoneNumber: "+36721601234567"
 });
 
 // Specify the maximum amount of time accepted