+++ examples/network-insights.ts
@@ -26,9 +26,9 @@ const myDevice = client.devices.get({
 // Set the duration of your subscription to congestion insights,
 // e.g.: it can end in `n` days starting from now.
 const tomorrowDate = new Date(Date.now() + 24 * 60 * 60 * 1000);
-tomorrowDate.setMilliseconds(0);
 
-// Subscribe your device to Congestion notifications
+// Subscribe your device to Congestion notifications FIRST.
+// Then, you'll be able to use other functionalities, such as polling.
 const congestionSubscription = await client.insights.subscribeToCongestionInfo(
     myDevice,
     tomorrowDate,
@@ -38,7 +38,7 @@ const congestionSubscription = await client.insights.subscribeToCongestionInfo(
 
 // Subscriptions are identified by id, for management
 // Use this to show the subscription:
-console.log(congestionSubscription.subscriptionId());
+console.log(congestionSubscription.subscriptionId);
 
 // Or check when your subscription starts/expires:
 console.log(congestionSubscription.startsAt);
@@ -46,6 +46,7 @@ console.log(congestionSubscription.expiresAt);
 
 // Get congestion predictions and historical data
 // between two timestamps with ISO 8601 formatted date strings.
+// It returns an array of congested objects and the prediction confidence level
 const congestion = await myDevice.getCongestion(
     // start date
     "2025-04-15T05:11:30.961136Z",