+++ integration-tests/qos.itest.ts
@@ -26,7 +26,7 @@ describe("Qos", () => {
                     privateAddress: "1.1.1.2",
                     publicPort: 80,
                 },
-                phoneNumber: `3670${
+                phoneNumber: `+3670${
                     Math.floor(Math.random() * (999999 - 123456 + 1)) + 123456
                 }`,
             });
@@ -37,7 +37,7 @@ describe("Qos", () => {
                     privateAddress: "1.1.1.2",
                     publicPort: 80,
                 },
-                phoneNumber: `3670${
+                phoneNumber: `+3670${
                     Math.floor(Math.random() * (999999 - 123456 + 1)) + 123456
                 }`,
             });
@@ -55,25 +55,35 @@ describe("Qos", () => {
 
     test("should create a session", async () => {
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
         expect(session.status).toEqual("REQUESTED");
         expect(session.profile).toEqual("QOS_L");
+        expect(session.serviceIpv4).toEqual("5.6.7.8");
+        expect(session.serviceIpv6).toEqual("2041:0000:140F::875B:131B");
         await session.deleteSession();
     });
 
     test("should create a session to a device with phone number", async () => {
         const session = await deviceWithPhoneNumber.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
         expect(session.status).toEqual("REQUESTED");
+        expect(session.device.phoneNumber).toEqual(
+            deviceWithPhoneNumber.phoneNumber
+        );
+        expect(session.serviceIpv4).toEqual("5.6.7.8");
+        expect(session.serviceIpv6).toEqual("2041:0000:140F::875B:131B");
         await session.deleteSession();
     });
 
     test("should create a session with medium profile", async () => {
         const session = await device.createQodSession("QOS_M", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
@@ -84,6 +94,7 @@ describe("Qos", () => {
 
     test("should create a session with small profile", async () => {
         const session = await device.createQodSession("QOS_S", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
@@ -94,6 +105,7 @@ describe("Qos", () => {
 
     test("should create a session with low latency profile", async () => {
         const session = await device.createQodSession("QOS_E", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
@@ -104,11 +116,15 @@ describe("Qos", () => {
 
     test("should get one session", async () => {
         const session = await device.createQodSession("QOS_E", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
         const fetchedSession = await client.sessions.get(session.id);
         expect(session.id).toEqual(fetchedSession.id);
+        expect(fetchedSession.device.networkAccessIdentifier).toEqual(
+            device.networkAccessIdentifier
+        );
         await session.deleteSession();
 
         try {
@@ -121,6 +137,7 @@ describe("Qos", () => {
 
     test("should get all sessions", async () => {
         await device.createQodSession("QOS_E", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
@@ -133,17 +150,20 @@ describe("Qos", () => {
 
     test("should create a session with service port", async () => {
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             servicePorts: { ports: [80] },
         });
         expect(session.status).toEqual("REQUESTED");
         expect(session.profile).toEqual("QOS_L");
+        expect(session.servicePorts?.ports).toEqual([80]);
         await session.deleteSession();
     });
 
     test("should create a session with service port range", async () => {
         const session = await device.createQodSession("QOS_M", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             servicePorts: { ranges: [{ from: 80, to: 443 }] },
@@ -151,11 +171,13 @@ describe("Qos", () => {
 
         expect(session.status).toEqual("REQUESTED");
         expect(session.profile).toEqual("QOS_M");
+        expect(session.servicePorts?.ranges).toEqual([{ from: 80, to: 443 }]);
         await session.deleteSession();
     });
 
     test("should create a session with device port", async () => {
         const session = await device.createQodSession("QOS_M", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             devicePorts: { ports: [80] },
@@ -168,6 +190,7 @@ describe("Qos", () => {
 
     test("should create a session with device port range", async () => {
         const session = await device.createQodSession("QOS_M", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             devicePorts: { ranges: [{ from: 80, to: 443 }] },
@@ -193,6 +216,7 @@ describe("Qos", () => {
 
     test("should create a session with notification url", async () => {
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             notificationAuthToken: "c8974e592c2fa383d4a3960714",
@@ -213,8 +237,13 @@ describe("Qos", () => {
             },
         });
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
         });
+
+        expect(session.serviceIpv4).toEqual("5.6.7.8");
+        expect(session.device.ipv4Address?.publicAddress).toEqual("1.1.1.2");
+        expect(session.device.ipv4Address?.privateAddress).toEqual("1.1.1.2");
 
         await session.deleteSession();
     });
@@ -228,8 +257,13 @@ describe("Qos", () => {
             },
         });
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
         });
+
+        expect(session.serviceIpv4).toEqual("5.6.7.8");
+        expect(session.device.ipv4Address?.publicAddress).toEqual("1.1.1.2");
+        expect(session.device.ipv4Address?.publicPort).toEqual(80);
 
         await session.deleteSession();
     });