+++ tests/qos.test.ts
@@ -46,6 +46,18 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                networkAccessIdentifier: "test-device@testcsp.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+            },
             qosStatus: "REQUESTED",
             startedAt: "2024-06-18T09:46:58.213Z",
             expiresAt: "2024-06-18T09:47:58.213Z",
@@ -65,6 +77,7 @@ describe("Qos", () => {
             applicationServer: {
                 ipv4Address: "5.6.7.8",
             },
+            duration: 3600,
         };
 
         fetchMock.post(
@@ -78,9 +91,15 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
         });
         expect(session.status).toEqual(mockResponse["qosStatus"]);
+        expect(session.device.networkAccessIdentifier).toEqual(
+            device.networkAccessIdentifier
+        );
+        expect(session.serviceIpv4).toEqual("5.6.7.8");
+        expect(session.serviceIpv6).toBeFalsy();
         expect(session.duration()).toEqual(60);
     });
 
@@ -98,6 +117,19 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -118,6 +150,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
         };
 
         fetchMock.post(
@@ -131,10 +164,13 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
         });
         expect(session.status).toEqual(mockResponse["qosStatus"]);
+        expect(session.device.ipv6Address).toEqual(device.ipv6Address);
+        expect(session.serviceIpv6).toEqual("2041:0000:140F::875B:131B");
     });
 
     test("should create a session with device ports", async () => {
@@ -152,6 +188,20 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -173,6 +223,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
             devicePorts: {
                 ports: [80, 3000],
             },
@@ -189,11 +240,14 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             devicePorts: { ports: [80, 3000] },
         });
         expect(session.status).toEqual(mockResponse["qosStatus"]);
+        expect(session.device.ipv6Address).toEqual(device.ipv6Address);
+        expect(session.serviceIpv6).toEqual("2041:0000:140F::875B:131B");
     });
 
     test("should create a session with device port range", async () => {
@@ -211,6 +265,20 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -232,6 +300,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
             devicePorts: {
                 ranges: [{ from: 80, to: 3000 }],
             },
@@ -248,6 +317,7 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             devicePorts: { ranges: [{ from: 80, to: 3000 }] },
@@ -269,6 +339,23 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
+            duration: 3600,
+            applicationServerPorts: {
+                ports: [80, 3000],
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -289,6 +376,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
             applicationServerPorts: {
                 ports: [80, 3000],
             },
@@ -305,11 +393,13 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             servicePorts: { ports: [80, 3000] },
         });
         expect(session.status).toEqual(mockResponse["qosStatus"]);
+        expect(session.servicePorts?.ports).toEqual([80, 3000]);
     });
 
     test("should create a session with service port range", async () => {
@@ -326,6 +416,24 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
+            duration: 3600,
+            applicationServerPorts: {
+                ranges: [{ from: 80, to: 3000 }],
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -347,6 +455,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
             applicationServerPorts: {
                 ranges: [{ from: 80, to: 3000 }],
             },
@@ -363,11 +472,13 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             servicePorts: { ranges: [{ from: 80, to: 3000 }] },
         });
         expect(session.status).toEqual(mockResponse["qosStatus"]);
+        expect(session.servicePorts?.ranges).toEqual([{ from: 80, to: 3000 }]);
     });
 
     test("should create a session with duration", async () => {
@@ -384,6 +495,20 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
             qosStatus: "REQUESTED",
             startedAt: "2024-06-18T09:46:58.213Z",
             expiresAt: "2024-06-18T09:47:58.213Z",
@@ -441,6 +566,20 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "08305343-7ed2-43b7-8eda-4c5ae9805bd0",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
+            applicationServer: {
+                ipv4Address: "5.6.7.8",
+                ipv6Address: "2041:0000:140F::875B:131B",
+            },
             qosStatus: "REQUESTED",
             startedAt: 1691671102,
             expiresAt: 1691757502,
@@ -462,6 +601,7 @@ describe("Qos", () => {
                 ipv4Address: "5.6.7.8",
                 ipv6Address: "2041:0000:140F::875B:131B",
             },
+            duration: 3600,
             notificationUrl: "https://example.com/notifications",
             notificationAuthToken: "Bearer c8974e592c2fa383d4a3960714",
         };
@@ -477,6 +617,7 @@ describe("Qos", () => {
         );
 
         const session = await device.createQodSession("QOS_L", {
+            duration: 3600,
             serviceIpv4: "5.6.7.8",
             serviceIpv6: "2041:0000:140F::875B:131B",
             notificationAuthToken: "c8974e592c2fa383d4a3960714",
@@ -489,6 +630,16 @@ describe("Qos", () => {
         let mockResponse = {
             sessionId: "1234",
             qosProfile: "QOS_L",
+            device: {
+                ipv4Address: {
+                    publicAddress: "1.1.1.2",
+                    privateAddress: "1.1.1.2",
+                    publicPort: 80,
+                },
+                ipv6Address: "2041:0000:140F::875B:131B",
+                networkAccessIdentifier: "testuser@open5glab.net",
+                phoneNumber: "9382948473",
+            },
             qosStatus: "BLA",
             expiresAt: 1641494400,
             startedAt: 0,
@@ -519,6 +670,16 @@ describe("Qos", () => {
             {
                 sessionId: "1234",
                 qosProfile: "QOS_L",
+                device: {
+                    ipv4Address: {
+                        publicAddress: "1.1.1.2",
+                        privateAddress: "1.1.1.2",
+                        publicPort: 80,
+                    },
+                    ipv6Address: "2041:0000:140F::875B:131B",
+                    networkAccessIdentifier: "testuser@open5glab.net",
+                    phoneNumber: "9382948473",
+                },
                 qosStatus: "BLA",
                 expiresAt: 1641494400,
                 startedAt: 0,
@@ -532,6 +693,10 @@ describe("Qos", () => {
 
         const sessions = await device.sessions();
         expect(sessions[0].id).toEqual("1234");
+        expect(sessions[0].device.networkAccessIdentifier).toEqual(
+            device.networkAccessIdentifier
+        );
+        expect(sessions[0].device.ipv6Address).toEqual(device.ipv6Address);
     });
 
     test("should not create a session without ip address", async () => {
@@ -567,5 +732,127 @@ describe("Qos", () => {
         } catch (error) {
             expect(true).toBe(true);
         }
+    });
+
+    test("should filter sessions by device network access identifier", async () => {
+        let device = client.devices.get({
+            networkAccessIdentifier: "testuser@open5glab.net",
+        });
+
+        let mockResponse = [
+            {
+                sessionId: "1234",
+                qosProfile: "QOS_L",
+                device: {
+                    networkAccessIdentifier: "testuser@open5glab.net",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+            {
+                sessionId: "12345",
+                qosProfile: "QOS_L",
+                device: {
+                    networkAccessIdentifier: "testuser2@open5glab.net",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+        ];
+
+        fetchMock.get(
+            "https://quality-of-service-on-demand.p-eu.rapidapi.com/sessions?networkAccessIdentifier=testuser@open5glab.net",
+            JSON.stringify(mockResponse)
+        );
+
+        const sessions = await device.sessions();
+        expect(sessions.length).toEqual(1);
+    });
+
+    test("should filter sessions by device phone number", async () => {
+        let device = client.devices.get({
+            phoneNumber: "+123214343",
+        });
+
+        let mockResponse = [
+            {
+                sessionId: "1234",
+                qosProfile: "QOS_L",
+                device: {
+                    phoneNumber: "+123214343",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+            {
+                sessionId: "12345",
+                qosProfile: "QOS_L",
+                device: {
+                    phoneNumber: "+321433443",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+            {
+                sessionId: "12346",
+                qosProfile: "QOS_L",
+                device: {
+                    phoneNumber: "+123214343",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+        ];
+
+        fetchMock.get(
+            "https://quality-of-service-on-demand.p-eu.rapidapi.com/sessions?phoneNumber=+123214343",
+            JSON.stringify(mockResponse)
+        );
+
+        const sessions = await device.sessions();
+        expect(sessions.length).toEqual(2);
+    });
+
+    test("should filter sessions by device phone number and ", async () => {
+        let device = client.devices.get({
+            networkAccessIdentifier: "testuser@open5glab.net",
+            phoneNumber: "+123214343",
+        });
+
+        let mockResponse = [
+            {
+                sessionId: "1234",
+                qosProfile: "QOS_L",
+                device: {
+                    networkAccessIdentifier: "testuser@open5glab.net",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+            {
+                sessionId: "12345",
+                qosProfile: "QOS_L",
+                device: {
+                    phoneNumber: "+321433443",
+                },
+                qosStatus: "BLA",
+                expiresAt: 1641494400,
+                startedAt: 0,
+            },
+        ];
+
+        fetchMock.get(
+            "https://quality-of-service-on-demand.p-eu.rapidapi.com/sessions?networkAccessIdentifier=testuser@open5glab.net",
+            JSON.stringify(mockResponse)
+        );
+
+        const sessions = await device.sessions();
+        expect(sessions.length).toEqual(1);
     });
 });